+++ test/rtos_drivers/usb/src/individual_tests/usb/local/usb_descriptors.c
@@ -23,11 +23,15 @@
  *
  */
 
+#include "tusb_config.h"
 #include "individual_tests/usb/local/usb_descriptors.h"
 #include "tusb.h"
+#include "class/dfu/dfu_device.h"
 
-#define XMOS_VID    0x20B1
-#define TEST_PID    0x4000
+#define XMOS_VID            0x20B1
+#define TEST_PID            0x4000
+#define USB_BCD             0x0200
+#define NUM_CONFIGURATIONS  0x01
 
 //--------------------------------------------------------------------+
 // Device Descriptors
@@ -35,11 +39,14 @@
 tusb_desc_device_t const desc_device = {
     .bLength            = sizeof(tusb_desc_device_t),
     .bDescriptorType    = TUSB_DESC_DEVICE,
-    .bcdUSB             = 0x0200,
+    .bcdUSB             = USB_BCD,
 
-    .bDeviceClass       = TUSB_CLASS_UNSPECIFIED,
-    .bDeviceSubClass    = TUSB_CLASS_UNSPECIFIED,
-    .bDeviceProtocol    = TUSB_CLASS_UNSPECIFIED,
+    // Use Interface Association Descriptor (IAD) for CDC
+    // As required by USB Specs IAD's subclass must be common class (2) and protocol must be IAD (1)
+    .bDeviceClass       = TUSB_CLASS_MISC,
+    .bDeviceSubClass    = MISC_SUBCLASS_COMMON,
+    .bDeviceProtocol    = MISC_PROTOCOL_IAD,
+
     .bMaxPacketSize0    = CFG_TUD_ENDPOINT0_SIZE,
 
     .idVendor           = XMOS_VID,
@@ -50,119 +57,102 @@ tusb_desc_device_t const desc_device = {
     .iProduct           = 0x02,
     .iSerialNumber      = 0x03,
 
-    .bNumConfigurations = 0x01
+    .bNumConfigurations = NUM_CONFIGURATIONS
 };
 
 // Invoked when received GET DEVICE DESCRIPTOR
 // Application return pointer to descriptor
 uint8_t const* tud_descriptor_device_cb(void)
 {
-    return (uint8_t const*) &desc_device;
+    return (uint8_t const *) &desc_device;
 }
 
 //--------------------------------------------------------------------+
 // Configuration Descriptor
 //--------------------------------------------------------------------+
 
-#define TUD_XMOS_DEVICE_CONTROL_DESC_LEN 9
+#define CONFIG_TOTAL_LEN        (TUD_CONFIG_DESC_LEN + CFG_TUD_CDC * TUD_CDC_DESC_LEN + TUD_DFU_DESC_LEN(ALT_COUNT))
 
-#define TUD_XMOS_DEVICE_CONTROL_DESCRIPTOR(_itfnum, _stridx) \
-  /* Interface */\
-  9, TUSB_DESC_INTERFACE, _itfnum, 0, 0, TUSB_CLASS_VENDOR_SPECIFIC, 0x00, 0x00, _stridx
+#define FUNC_ATTRS              (DFU_ATTR_CAN_UPLOAD | DFU_ATTR_CAN_DOWNLOAD | DFU_ATTR_WILL_DETACH | DFU_ATTR_MANIFESTATION_TOLERANT)
 
-const size_t uac2_interface_descriptors_length =
-        TUD_AUDIO_DESC_CLK_SRC_LEN +
-        + TUD_AUDIO_DESC_INPUT_TERM_LEN
-        + TUD_AUDIO_DESC_OUTPUT_TERM_LEN
-        + TUD_AUDIO_DESC_INPUT_TERM_LEN
-        + TUD_AUDIO_DESC_OUTPUT_TERM_LEN
-        ;
+#define EP_NUM_CDC_0_NOTIF      0x81
+#define EP_NUM_CDC_0_OUT        0x02
+#define EP_NUM_CDC_0_IN         0x82
 
-const size_t uac2_total_descriptors_length =
-        TUD_AUDIO_DESC_IAD_LEN +
-        TUD_AUDIO_DESC_STD_AC_LEN +
-        TUD_AUDIO_DESC_CS_AC_LEN +
-        uac2_interface_descriptors_length +
-        + TUD_AUDIO_DESC_STD_AS_INT_LEN
-        + TUD_AUDIO_DESC_STD_AS_INT_LEN
-        + TUD_AUDIO_DESC_CS_AS_INT_LEN
-        + TUD_AUDIO_DESC_TYPE_I_FORMAT_LEN
-        + TUD_AUDIO_DESC_STD_AS_ISO_EP_LEN
-        + TUD_AUDIO_DESC_CS_AS_ISO_EP_LEN
-        + TUD_AUDIO_DESC_STD_AS_INT_LEN
-        + TUD_AUDIO_DESC_STD_AS_INT_LEN
-        + TUD_AUDIO_DESC_CS_AS_INT_LEN
-        + TUD_AUDIO_DESC_TYPE_I_FORMAT_LEN
-        + TUD_AUDIO_DESC_STD_AS_ISO_EP_LEN
-        + TUD_AUDIO_DESC_CS_AS_ISO_EP_LEN
-        ;
+#define EP_NUM_CDC_1_NOTIF      0x83
+#define EP_NUM_CDC_1_OUT        0x04
+#define EP_NUM_CDC_1_IN         0x84
 
-// List of audio descriptor lengths which is required by audio driver - you need as many entries as CFG_TUD_AUDIO - unfortunately this is not possible to determine otherwise
-const uint16_t tud_audio_desc_lengths[CFG_TUD_AUDIO] = {
-        uac2_total_descriptors_length
-};
-
-#define CONFIG_TOTAL_LEN        (TUD_CONFIG_DESC_LEN + CFG_TUD_AUDIO * uac2_total_descriptors_length)
-#define EPNUM_AUDIO   0x01
-
-// TAKE CARE - THE NUMBER OF AUDIO STREAMING INTERFACES PER AUDIO FUNCTION MUST NOT EXCEED CFG_TUD_AUDIO_N_AS_INT - IF IT DOES INCREASE CFG_TUD_AUDIO_N_AS_INT IN tusb_config.h!
-
-#define AUDIO_INTERFACE_STRING_INDEX 4
-
-uint8_t const desc_configuration[] = {
+uint8_t const desc_fs_configuration[] = {
     // Interface count, string index, total length, attribute, power in mA
-    TUD_CONFIG_DESCRIPTOR(1, ITF_NUM_TOTAL, 0, CONFIG_TOTAL_LEN, TUSB_DESC_CONFIG_ATT_REMOTE_WAKEUP, 400),
+    TUD_CONFIG_DESCRIPTOR(1, ITF_NUM_TOTAL, 0, CONFIG_TOTAL_LEN, 0x00, 500),
 
-    /* Standard Interface Association Descriptor (IAD) */
-    TUD_AUDIO_DESC_IAD(/*_firstitfs*/ ITF_NUM_AUDIO_CONTROL, /*_nitfs*/ 3, /*_stridx*/ 0x00),
-    /* Standard AC Interface Descriptor(4.7.1) */
-    TUD_AUDIO_DESC_STD_AC(/*_itfnum*/ ITF_NUM_AUDIO_CONTROL, /*_nEPs*/ 0x00, /*_stridx*/ AUDIO_INTERFACE_STRING_INDEX),
-    /* Class-Specific AC Interface Header Descriptor(4.7.2) */
-    TUD_AUDIO_DESC_CS_AC(/*_bcdADC*/ 0x0200, /*_category*/ AUDIO_FUNC_OTHER, /*_totallen*/ uac2_interface_descriptors_length, /*_ctrl*/ AUDIO_CS_AS_INTERFACE_CTRL_LATENCY_POS),
-    /* Clock Source Descriptor(4.7.2.1) */
-    TUD_AUDIO_DESC_CLK_SRC(/*_clkid*/ UAC2_ENTITY_CLOCK, /*_attr*/ AUDIO_CLOCK_SOURCE_ATT_INT_PRO_CLK, /*_ctrl*/ (AUDIO_CTRL_R << AUDIO_CLOCK_SOURCE_CTRL_CLK_VAL_POS) | (AUDIO_CTRL_R << AUDIO_CLOCK_SOURCE_CTRL_CLK_FRQ_POS), /*_assocTerm*/ 0x00,  /*_stridx*/ 0x00),
+    // 1st CDC: Interface number, string index, EP notification address and size, EP data address (out, in) and size.
+    TUD_CDC_DESCRIPTOR(ITF_NUM_CDC_0, 4, EP_NUM_CDC_0_NOTIF, 8, EP_NUM_CDC_0_OUT, EP_NUM_CDC_0_IN, 64),
 
-    /* Input Terminal Descriptor(4.7.2.4) */
-    TUD_AUDIO_DESC_INPUT_TERM(/*_termid*/ UAC2_ENTITY_SPK_INPUT_TERMINAL, /*_termtype*/ AUDIO_TERM_TYPE_USB_STREAMING, /*_assocTerm*/ 0x00, /*_clkid*/ UAC2_ENTITY_CLOCK, /*_nchannelslogical*/ CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_RX, /*_channelcfg*/ AUDIO_CHANNEL_CONFIG_NON_PREDEFINED, /*_idxchannelnames*/ 0x00, /*_ctrl*/ AUDIO_CTRL_NONE, /*_stridx*/ 0x00),
-    /* Output Terminal Descriptor(4.7.2.5) */
-    TUD_AUDIO_DESC_OUTPUT_TERM(/*_termid*/ UAC2_ENTITY_SPK_OUTPUT_TERMINAL, /*_termtype*/ AUDIO_TERM_TYPE_OUT_GENERIC_SPEAKER, /*_assocTerm*/ 0x00, /*_srcid*/ UAC2_ENTITY_SPK_INPUT_TERMINAL, /*_clkid*/ UAC2_ENTITY_CLOCK, /*_ctrl*/ AUDIO_CTRL_NONE, /*_stridx*/ 0x00),
+    // 2nd CDC: Interface number, string index, EP notification address and size, EP data address (out, in) and size.
+    TUD_CDC_DESCRIPTOR(ITF_NUM_CDC_1, 4, EP_NUM_CDC_1_NOTIF, 8, EP_NUM_CDC_1_OUT, EP_NUM_CDC_1_IN, 64),
 
-    /* Input Terminal Descriptor(4.7.2.4) */
-    TUD_AUDIO_DESC_INPUT_TERM(/*_termid*/ UAC2_ENTITY_MIC_INPUT_TERMINAL, /*_termtype*/ AUDIO_TERM_TYPE_IN_GENERIC_MIC, /*_assocTerm*/ 0x00, /*_clkid*/ UAC2_ENTITY_CLOCK, /*_nchannelslogical*/ CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_TX, /*_channelcfg*/ AUDIO_CHANNEL_CONFIG_NON_PREDEFINED, /*_idxchannelnames*/ 0x00, /*_ctrl*/ AUDIO_CTRL_NONE, /*_stridx*/ 0x00),
-
-    /* Output Terminal Descriptor(4.7.2.5) */
-    TUD_AUDIO_DESC_OUTPUT_TERM(/*_termid*/ UAC2_ENTITY_MIC_OUTPUT_TERMINAL, /*_termtype*/ AUDIO_TERM_TYPE_USB_STREAMING, /*_assocTerm*/ 0x00, /*_srcid*/ UAC2_ENTITY_MIC_INPUT_TERMINAL, /*_clkid*/ UAC2_ENTITY_CLOCK, /*_ctrl*/ AUDIO_CTRL_NONE, /*_stridx*/ 0x00),
-
-    /* Standard AS Interface Descriptor(4.9.1) */
-    /* Interface 1, Alternate 0 - default alternate setting with 0 bandwidth */
-    TUD_AUDIO_DESC_STD_AS_INT(/*_itfnum*/ ITF_NUM_AUDIO_STREAMING_SPK, /*_altset*/ 0x00, /*_nEPs*/ 0x00, /*_stridx*/ 0x00),
-    /* Standard AS Interface Descriptor(4.9.1) */
-    /* Interface 1, Alternate 1 - alternate interface for data streaming */
-    TUD_AUDIO_DESC_STD_AS_INT(/*_itfnum*/ ITF_NUM_AUDIO_STREAMING_SPK, /*_altset*/ 0x01, /*_nEPs*/ 0x01, /*_stridx*/ 0x00),
-    /* Class-Specific AS Interface Descriptor(4.9.2) */
-    TUD_AUDIO_DESC_CS_AS_INT(/*_termid*/ UAC2_ENTITY_SPK_INPUT_TERMINAL, /*_ctrl*/ AUDIO_CTRL_NONE, /*_formattype*/ AUDIO_FORMAT_TYPE_I, /*_formats*/ AUDIO_DATA_FORMAT_TYPE_I_PCM, /*_nchannelsphysical*/ CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_RX, /*_channelcfg*/ AUDIO_CHANNEL_CONFIG_NON_PREDEFINED, /*_stridx*/ 0x00),
-    /* Type I Format Type Descriptor(2.3.1.6 - Audio Formats) */
-    TUD_AUDIO_DESC_TYPE_I_FORMAT(CFG_TUD_AUDIO_FUNC_1_N_BYTES_PER_SAMPLE_RX, CFG_TUD_AUDIO_FUNC_1_N_BYTES_PER_SAMPLE_RX*8),
-    /* Standard AS Isochronous Audio Data Endpoint Descriptor(4.10.1.1) */
-    TUD_AUDIO_DESC_STD_AS_ISO_EP(/*_ep*/ EPNUM_AUDIO, /*_attr*/ (TUSB_XFER_ISOCHRONOUS | TUSB_ISO_EP_ATT_ADAPTIVE | /*TUSB_ISO_EP_ATT_IMPLICIT_FB |*/ TUSB_ISO_EP_ATT_DATA), /*_maxEPsize*/ CFG_TUD_AUDIO_FUNC_1_EP_OUT_SZ, /*_interval*/ (CFG_TUSB_RHPORT0_MODE & OPT_MODE_HIGH_SPEED) ? 0x04 : 0x01),
-    /* Class-Specific AS Isochronous Audio Data Endpoint Descriptor(4.10.1.2) */
-    TUD_AUDIO_DESC_CS_AS_ISO_EP(/*_attr*/ AUDIO_CS_AS_ISO_DATA_EP_ATT_NON_MAX_PACKETS_OK, /*_ctrl*/ AUDIO_CTRL_NONE, /*_lockdelayunit*/ AUDIO_CS_AS_ISO_DATA_EP_LOCK_DELAY_UNIT_UNDEFINED, /*_lockdelay*/ 0x0000),
-
-    /* Standard AS Interface Descriptor(4.9.1) */
-    /* Interface 1, Alternate 0 - default alternate setting with 0 bandwidth */
-    TUD_AUDIO_DESC_STD_AS_INT(/*_itfnum*/ ITF_NUM_AUDIO_STREAMING_MIC, /*_altset*/ 0x00, /*_nEPs*/ 0x00, /*_stridx*/ 0x00),
-    /* Standard AS Interface Descriptor(4.9.1) */
-    /* Interface 1, Alternate 1 - alternate interface for data streaming */
-    TUD_AUDIO_DESC_STD_AS_INT(/*_itfnum*/ ITF_NUM_AUDIO_STREAMING_MIC, /*_altset*/ 0x01, /*_nEPs*/ 0x01, /*_stridx*/ 0x00),
-    /* Class-Specific AS Interface Descriptor(4.9.2) */
-    TUD_AUDIO_DESC_CS_AS_INT(/*_termid*/ UAC2_ENTITY_MIC_OUTPUT_TERMINAL, /*_ctrl*/ AUDIO_CTRL_NONE, /*_formattype*/ AUDIO_FORMAT_TYPE_I, /*_formats*/ AUDIO_DATA_FORMAT_TYPE_I_PCM, /*_nchannelsphysical*/ CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_TX, /*_channelcfg*/ AUDIO_CHANNEL_CONFIG_NON_PREDEFINED, /*_stridx*/ 0x00),
-    /* Type I Format Type Descriptor(2.3.1.6 - Audio Formats) */
-    TUD_AUDIO_DESC_TYPE_I_FORMAT(CFG_TUD_AUDIO_FUNC_1_N_BYTES_PER_SAMPLE_TX, CFG_TUD_AUDIO_FUNC_1_N_BYTES_PER_SAMPLE_TX*8),
-    /* Standard AS Isochronous Audio Data Endpoint Descriptor(4.10.1.1) */
-    TUD_AUDIO_DESC_STD_AS_ISO_EP(/*_ep*/ 0x80 | EPNUM_AUDIO, /*_attr*/ (TUSB_XFER_ISOCHRONOUS | TUSB_ISO_EP_ATT_ADAPTIVE | /*TUSB_ISO_EP_ATT_IMPLICIT_FB |*/ TUSB_ISO_EP_ATT_DATA), /*_maxEPsize*/ CFG_TUD_AUDIO_FUNC_1_EP_IN_SZ, /*_interval*/ (CFG_TUSB_RHPORT0_MODE & OPT_MODE_HIGH_SPEED) ? 0x04 : 0x01),
-    /* Class-Specific AS Isochronous Audio Data Endpoint Descriptor(4.10.1.2) */
-    TUD_AUDIO_DESC_CS_AS_ISO_EP(/*_attr*/ AUDIO_CS_AS_ISO_DATA_EP_ATT_NON_MAX_PACKETS_OK, /*_ctrl*/ AUDIO_CTRL_NONE, /*_lockdelayunit*/ AUDIO_CS_AS_ISO_DATA_EP_LOCK_DELAY_UNIT_UNDEFINED, /*_lockdelay*/ 0x0000),
+    // Interface number, Alternate count, starting string index, attributes, detach timeout, transfer size
+    TUD_DFU_DESCRIPTOR(ITF_NUM_DFU_MODE, ALT_COUNT, 5, FUNC_ATTRS, 1000, CFG_TUD_DFU_XFER_BUFSIZE)
 };
+
+#if TUD_OPT_HIGH_SPEED
+// Per USB specs: high speed capable device must report device_qualifier and other_speed_configuration
+
+uint8_t const desc_hs_configuration[] =
+{
+    // Config number, interface count, string index, total length, attribute, power in mA
+    TUD_CONFIG_DESCRIPTOR(1, ITF_NUM_TOTAL, 0, CONFIG_TOTAL_LEN, 0x00, 500),
+
+    // 1st CDC: Interface number, string index, EP notification address and size, EP data address (out, in) and size.
+    TUD_CDC_DESCRIPTOR(ITF_NUM_CDC_0, 4, EP_NUM_CDC_0_NOTIF, 8, EP_NUM_CDC_0_OUT, EP_NUM_CDC_0_IN, 512),
+
+    // 2nd CDC: Interface number, string index, EP notification address and size, EP data address (out, in) and size.
+    TUD_CDC_DESCRIPTOR(ITF_NUM_CDC_1, 4, EP_NUM_CDC_1_NOTIF, 8, EP_NUM_CDC_1_OUT, EP_NUM_CDC_1_IN, 512),
+
+    // Interface number, Alternate count, starting string index, attributes, detach timeout, transfer size
+    TUD_DFU_DESCRIPTOR(ITF_NUM_DFU_MODE, ALT_COUNT, 5, FUNC_ATTRS, 1000, CFG_TUD_DFU_XFER_BUFSIZE)
+};
+
+// device qualifier is mostly similar to device descriptor since we don't change configuration based on speed
+tusb_desc_device_qualifier_t const desc_device_qualifier =
+{
+    .bLength            = sizeof(tusb_desc_device_t),
+    .bDescriptorType    = TUSB_DESC_DEVICE,
+    .bcdUSB             = USB_BCD,
+
+    .bDeviceClass       = TUSB_CLASS_MISC,
+    .bDeviceSubClass    = MISC_SUBCLASS_COMMON,
+    .bDeviceProtocol    = MISC_PROTOCOL_IAD,
+
+    .bMaxPacketSize0    = CFG_TUD_ENDPOINT0_SIZE,
+    .bNumConfigurations = NUM_CONFIGURATIONS,
+    .bReserved          = 0x00
+};
+
+// Invoked when received GET DEVICE QUALIFIER DESCRIPTOR request
+// Application return pointer to descriptor, whose contents must exist long enough for transfer to complete.
+// device_qualifier descriptor describes information about a high-speed capable device that would
+// change if the device were operating at the other speed. If not highspeed capable stall this request.
+uint8_t const* tud_descriptor_device_qualifier_cb(void)
+{
+    return (uint8_t const*) &desc_device_qualifier;
+}
+
+// Invoked when received GET OTHER SEED CONFIGURATION DESCRIPTOR request
+// Application return pointer to descriptor, whose contents must exist long enough for transfer to complete
+// Configuration descriptor in the other speed e.g if high speed then this is for full speed and vice versa
+uint8_t const* tud_descriptor_other_speed_configuration_cb(uint8_t index)
+{
+    (void) index; // for multiple configurations
+
+    // if link speed is high return fullspeed config, and vice versa
+    return (tud_speed_get() == TUSB_SPEED_HIGH) ?  desc_fs_configuration : desc_hs_configuration;
+}
+
+#endif // TUD_OPT_HIGH_SPEED
+
 
 // Invoked when received GET CONFIGURATION DESCRIPTOR
 // Application return pointer to descriptor
@@ -170,7 +160,13 @@ uint8_t const desc_configuration[] = {
 uint8_t const* tud_descriptor_configuration_cb(uint8_t index)
 {
     (void) index; // for multiple configurations
-    return desc_configuration;
+
+#if TUD_OPT_HIGH_SPEED
+    // Although we are highspeed, host may be fullspeed.
+    return (tud_speed_get() == TUSB_SPEED_HIGH) ?  desc_hs_configuration : desc_fs_configuration;
+#else
+    return desc_fs_configuration;
+#endif
 }
 
 //--------------------------------------------------------------------+
@@ -178,12 +174,16 @@ uint8_t const* tud_descriptor_configuration_cb(uint8_t index)
 //--------------------------------------------------------------------+
 
 // array of pointer to string descriptors
-char const *string_desc_arr[] = {(const char[]) {0x09, 0x04}, // 0: is supported language is English (0x0409)
-        "XMOS",                    // 1: Manufacturer
-        "DRIVERTEST",              // 2: Product
-        "123456",                  // 3: Serials
-        "UAC2",                    // 4: Audio Interface
-        };
+char const *string_desc_arr[] = {
+    (const char[]) {0x09, 0x04}, // 0: is supported language is English (0x0409)
+    "XMOS",                      // 1: Manufacturer
+    "DRIVERTEST",                // 2: Product
+    "123456",                    // 3: Serials
+    "TEST CDC",                  // 4: CDC Interface
+    "TEST PARTITION 1",          // 5: DFU device
+    "TEST PARTITION 2",          // 6: DFU device
+    "TEST PARTITION 3",          // 7: DFU device
+};
 
 static uint16_t _desc_str[32];