+++ CHANGELOG.rst
@@ -1,12 +1,14 @@
 RTOS Framework change log
 =========================
 
-TBD
+3.0.0
 -----
   
-  * REMOVED: qspi_io based RTOS flash driver
+  * REMOVED: QSPI IO based RTOS flash driver
   * ADDED: lib_qspi_fast_read and lib_quadflash based RTOS flash driver
   * ADDED: RTOS flash driver function to set core affinity for QSPI transactions
+  * ADDED: New API to specify raw or nibble swapped individual QSPI flash reads
+  * UPDATED: Mbed TLS to version 2.28.3
 
 2.0.0
 -----